backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "logfire[fastapi]>=4.25.0",
    "mcp-hackernews>=1.1.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.62.0",
    "pydantic-ai-slim[duckduckgo]>=1.62.0",
    "pydantic-ai-slim[mcp]>=1.62.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
